requirements.txt
streamlit==1.28.0
facebook-business==18.0.0
google-generativeai==0.3.1
reportlab==4.0.7
weasyprint==60.1
plotly==5.17.0
pandas==2.1.3
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
